+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The official release will have a defined and more stable API. If you are already
  * Allow retriggering of nested steps (#5)
  * Redirect to new build when retriggering
  * Support hardcoded result-maps in `chain`-macro
+ * Support setting environment-variables in `shell/bash` (#9)
 * API Changes:
  * removed deprecated argument lists for `lambdacd.internal.execution/execute-steps`